switch from u/long to u32/s32

+++ bfsmacklib/include/bfsmack.h
@@ -37,12 +37,12 @@ enum SmackerPlayFlags {
     SMK_UnknFlag100        = 0x0100,
 };
 
-typedef void (*SmackDrawCallback)(ubyte *frame_data, long width, long height);
+typedef void (*SmackDrawCallback)(ubyte *frame_data, s32 width, s32 height);
 
 /******************************************************************************/
 extern SmackDrawCallback smack_draw_callback;
 
-TbResult play_smk(const char *fname, ulong smkflags, ushort plyflags);
+TbResult play_smk(const char *fname, u32 smkflags, ushort plyflags);
 
 void set_smack_malloc(void *(*cb)(int));
 void set_smack_free(void (*cb)(void *ptr));
