+++ bfsoundlib/include/bfaudio.h
@@ -60,8 +60,8 @@ void EnsureAILStartup(void);
 void InitAudio(AudioInitOptions *audOpts);
 void FreeAudio(void);
 
-void SetSoundMasterVolume(long vol);
-void SetMusicMasterVolume(long vol);
+void SetSoundMasterVolume(s32 vol);
+void SetMusicMasterVolume(s32 vol);
 
 /******************************************************************************/
 #ifdef __cplusplus
