+++ bfsoundlib/include/ssampfad.h
@@ -28,7 +28,7 @@ extern "C" {
 #endif
 /******************************************************************************/
 
-void FadeSample(long source_id, short smp_id, ushort to_volume,
+void FadeSample(s32 source_id, short smp_id, ushort to_volume,
   ubyte step, ubyte stop);
 
 void StopAllSampleFadeTimers(void);
