+++ bfsoundlib/src/init_snd.c
@@ -39,21 +39,21 @@
 #pragma pack(1)
 
 struct SoundBankSizes {
-    long DatSize;
-    long TabSize;
+    s32 DatSize;
+    s32 TabSize;
 };
 
 #pragma pack()
 
 TbBool SixteenBit = true;
-ulong SampleRate = 0;
+u32 SampleRate = 0;
 
 void *Sfx;
 void *EndSfxs;
 void *SfxData;
 
-long largest_dat_size = 0;
-long largest_tab_size = 0;
+s32 largest_dat_size = 0;
+s32 largest_tab_size = 0;
 char full_sound_data_path[224];
 struct SoundBankSizes sound_bank_size_info[9];
 
@@ -69,9 +69,9 @@ extern TbBool SoundActive;
 extern ushort SoundType;
 extern char SoundDataPath[144];
 
-extern ulong MaxNumberOfSamples;
+extern u32 MaxNumberOfSamples;
 extern TbBool StereoSound;
-extern long CurrentSoundMasterVolume;
+extern s32 CurrentSoundMasterVolume;
 
 extern struct SampleInfo sample_id[32];
 extern struct SampleInfo *end_sample_id;
@@ -172,9 +172,9 @@ int SetSoundBitsAndRate(ushort snd_type)
 int AllocateSoundBankMemory(ushort snd_type)
 {
     TbFileHandle fh;
-    long len;
+    s32 len;
     ushort tpno;
-    long dat_size, tab_size;
+    s32 dat_size, tab_size;
 
     sprintf(full_sound_data_path, "%s/sound.dat", SoundDataPath);
     fh = LbFileOpen(full_sound_data_path, Lb_FILE_MODE_READ_ONLY);
